requirements.txt
Flask
Werkzeug
yara-python
